site/requirements/base.txt
# -------------------------------------
# BASE REQUIREMENTS
# -------------------------------------

# Django Reqs
# =====================================

Django~=1.11
psycopg2-binary~=2.7
mimeparse
pytz

# Django Apps
# ======================================

django-robots
django-environ
django-extensions
django-storages
django-model-utils
djangorestframework
django-ipware
django-crispy-forms
django-admin-sortable2

# Project Reqs
# =====================================

pytest-django
pytest-mock
pytest-pythonpath
factory_boy
Faker
whitenoise
requests
dicttoxml
pyuploadcare
Pillow==5.3.0
django-solo==1.1.3
django-optimized-image~=0.2.0
